app/requirements.txt
asgiref==3.2.10
astroid==2.4.2
Django==3.1
django-markdownx==3.0.1
importlib-metadata==1.7.0
isort==5.4.2
lazy-object-proxy==1.4.3
Markdown==3.2.2
mccabe==0.6.1
Pillow==7.2.0
pylint==2.6.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
wrapt==1.12.1
zipp==3.1.0
psycopg2-binary==2.8.5
gunicorn==20.0.4
